parser: introduce basic AST and parser
This is quasi-initial commit of the DOOML language compiler done in
terms of functional language compilers 2025 SPbU course.

Basically, it is OCaml-like language with (drastically) truncated list
of features. This patch introduces a basic AST and parser with
pretty-printing facilities. Only integers, units and tuples are
supported yet.

This will be implemented by Georgiy Belyanin and Ignatiy Sergeev.