me: add a-normal-form
This patch introduces a-normal-form (ANF) as a step in the compiler's
middle-end. ANF is responsible to make the function calls only contain
variables or immediate values as their arguments which makes
the code kind of resemble assembly in the sense that assembler
instructions usually work with registers/immediate values not with
compound statements.

Example:

```ocaml
(* Before ANF *)
let f =
  let q = f ((g + sup0) * (2 * i)) in
  q
;;

(* After ANF *)
let f =
  let sup2 = (*) 2 i in
  let sup5 = (+) g sup0 in
  let sup6 = (*) sup5 sup2 in
  let sup7 = (f) sup6 in
  let q = sup7 in
  q
;;
```